Move admin action types to main types file
